+++ desktop/php/Abeille-Eq-Advanced-Common.php
@@ -9,59 +9,35 @@
 
 <div class="form-group">
     <label class="col-sm-3 control-label">{{Nom}}</label>
-    <div class="col-sm-5">
-    <!-- <span>< ?php echo $eqName; ?></span> -->
-    <input id="idEqName" type="text" value="" readonly>
+    <div class="col-sm-9">
+        <input id="idEqName" type="text" value="" readonly style="width:240px">
     </div>
 </div>
 
 <div class="form-group">
     <label class="col-sm-3 control-label">{{ID Jeedom}}</label>
-    <div class="col-sm-5">
-        <input id="idEqId" type="text" value="" readonly>
+    <div class="col-sm-9">
+        <input id="idEqId" type="text" value="" readonly style="width:240px">
     </div>
 </div>
 
 <div class="form-group">
     <label class="col-sm-3 control-label">{{Nom logique}}</label>
     <div class="col-sm-5">
-        <input type="text" class="eqLogicAttr" data-l1key="logicalId"></input>
+        <input type="text" class="eqLogicAttr" data-l1key="logicalId" style="width:240px"></input>
     </div>
 </div>
 
 <div class="form-group">
     <label class="col-sm-3 control-label">{{Dernière comm.}}</label>
-    <div advInfo="Time-Time" class="col-sm-5">
-        <input type="text" value="" readonly>
+    <div advInfo="Time-Time" class="col-sm-9">
+        <input type="text" value="" readonly style="width:240px">
     </div>
 </div>
-
-<!-- <div class="form-group">
-    <label class="col-sm-3 control-label">{{Adresse (courte/IEEE)}}</label>
-    <div class="col-sm-5">
-        <input id="idEqAddr" type="text" value="" readonly>
-        /
-        <span id="idEqIeee" class="eqLogicAttr" data-l1key="configuration" data-l2key="IEEE"></span>
-    </div>
-</div> -->
 
 <div class="form-group">
     <label class="col-sm-3 control-label">{{Documentation}}</label>
-    <div class="col-sm-5">
+    <div class="col-sm-9">
         <a id="idDocUrl" href="tobefilled" target="_blank">{{Voir ici si présente}}</a>
     </div>
 </div>
-
-<!-- <div class="form-group">
-    <label class="col-sm-3 control-label" title="{{Groupes Zigbee auquels l'équipement appartient}}">{{Groupes}}</label>
-    <div class="col-sm-5">
-        < ?php
-            $cmd = AbeilleCmd::byEqLogicIdAndLogicalId($eqId, 'Group-Membership');
-            if (is_object($cmd))
-                $groups = $cmd->execCmd();
-            else
-                $groups = "";
-            echo '<span>'.$groups.'</span>';
-        ?>
-    </div>
-</div> -->